requirements.txt
# Core
Flask==3.1.1
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1

# Database
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0

# Forms and Validation  
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0.post1

# Development
python-dotenv==1.0.0
blinker==1.9.0

# Testing
pytest==8.1.1
pytest-cov==4.1.0
